requirements.txt
dash==1.12.0
dash-bootstrap-components==0.10.2
dash-core-components==1.10.0
dash-cytoscape==0.1.1
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
gensim==3.8.2
numpy==1.17.5
pandas==0.25.1
pdfminer.six==20200517
plotly==4.6.0
python==3.7.2
python-docx==0.8.10
pip==20.1.1
scikit-learn==0.22.1
